Library (error.h): high-level functions for simple calling
Until now, each error code in each library was manually given an interger
idnetifier. But the manual nature of it can cause many problems in the
future (with typos for example). However, since the error code in the
structure is defined as a single integer with three bit-wise components, it
was complicated to edit.

On the other hand, with this new error structure, checking and aborting the
error status inside Gnuastro's programs was taking too many lines of code
and becoming error-prone.

With this commit, each library now has a 'LIBRARY_error()' function which
takes the specific error code, and adds the library code for constructing
the full error code. Using this high-level function we can now use an
'enum' for generating the error code values of each situation.

Furthermore, to address the issue with checking and aborting the programs
in case of an error, we defined a new internal header file which contains
the 'exit(EXIT_FAILURE)' command. This special internal header should only
be called insdie the programs, not the library. The new file was also added
in the 'EXTRA_DIST' variable of 'lib/Makefile.am'.

Some other minor issues have been implemented:

 - The 'gnuastro/error.h' is now included in all the library headers
   because it will be necessary as an argument type.

 - Also, to simplify checking for an error at the start of each function a
   new 'gal_error_func_reject' has been added.