Library (error.h): apparent bug
Until now, the numbers given to arguments when printing an error message
are not printed in the output string.

With this commit, to reproduce the problem, some checks within
CosmicCalculator have been disabled.

Later, when we find the cause of the problem, we should remove all the
changes in this commit.